requirements.txt
numpy
opencv-python
matplotlib
Django==5.0.6
django-q2
GDAL
pillow
scikit-learn==1.3.1
scipy=1.13.1
redis
pyproj